+++ beaker/exceptions.py
@@ -165,3 +165,9 @@ class DuplicateExperimentError(BeakerError):
 
 class TaskNotFound(BeakerError):
     pass
+
+
+class ChecksumFailedError(BeakerError):
+    """
+    Raised when a downloaded file from a Beaker dataset is corrupted.
+    """